Changed the datatype for 'episode' in 'appears_in' from INTEGER to SMALLINT, as this causes problems with Postgres and deserialization

+++ wundergraph_example/migrations/pg/2018-01-24-131925_setup/up.sql
@@ -20,7 +20,7 @@ CREATE TABLE heros(
 
 CREATE TABLE appears_in(
     hero_id INTEGER NOT NULL REFERENCES heros(id) ON DELETE CASCADE ON UPDATE RESTRICT,
-    episode INTEGER NOT NULL CHECK(episode IN (1,2,3)),
+    episode SMALLINT NOT NULL CHECK(episode IN (1,2,3)),
     PRIMARY KEY(hero_id, episode)
 );
